test(app): isolate session files rail proof
