+++ evals/specs/artifact-rail-layout.test.ts
@@ -7,10 +7,6 @@ const sessionPageSource = readFileSync(
   fileURLToPath(new URL("../../apps/app/src/react-app/domains/session/chat/session-page.tsx", import.meta.url)),
   "utf8",
 );
-const messageListSource = readFileSync(
-  fileURLToPath(new URL("../../apps/app/src/components/chat/message-list.tsx", import.meta.url)),
-  "utf8",
-);
 const buttonSource = readFileSync(
   fileURLToPath(new URL("../../apps/app/src/components/ui/button.tsx", import.meta.url)),
   "utf8",
@@ -60,20 +56,6 @@ test("the artifact count badge stays inside the desktop rail", async ({ evidence
   evidence.recordAssertionEvidence(
     "Artifact counts remain fully visible in the desktop rail",
     "The one-digit and compact 9+ badges fit inside the rail at narrow and wide chat widths without changing the 9+ presentation.",
-    true,
-  );
-});
-
-test("session files stay in the side rail instead of the chat transcript", async ({ evidence }) => {
-  expect(messageListSource).not.toContain("ArtifactList");
-  expect(messageListSource).not.toContain("data-files-strip");
-  expect(sessionPageSource).toContain('title={`Files (${artifactTargetCount})`}');
-  expect(sessionPageSource).toContain('aria-label={`Files (${artifactTargetCount})`}');
-  expect(sessionPageSource).toContain("onClick={openArtifactRailPane}");
-
-  evidence.recordAssertionEvidence(
-    "Session files are accessed from the side rail, not repeated in the transcript",
-    "The message list has no per-turn files strip, while the session side rail keeps a labeled Files control that opens the artifact pane.",
     true,
   );
 });
